+++ crates/brynja-mac-kmac/src/backend.rs
@@ -11,7 +11,7 @@ pub(crate) trait CshakeState: Sized {
     fn check_additional_bytes(&self, additional: u128) -> Result<(), HardenedSha3Error>;
     fn update(&mut self, input: &[u8]) -> Result<(), HardenedSha3Error>;
     fn wipe_in_place(&mut self);
-    fn finalize_xof_erasing_source(&mut self) -> Self::Reader;
+    fn finalize_xof_erasing_source(&mut self) -> Result<Self::Reader, HardenedSha3Error>;
     fn finalize_bits_xof_erasing_source(
         &mut self,
         input: Fips202BitString<'_>,
@@ -57,7 +57,7 @@ macro_rules! backend {
                 Self::wipe_in_place(self);
             }
 
-            fn finalize_xof_erasing_source(&mut self) -> Self::Reader {
+            fn finalize_xof_erasing_source(&mut self) -> Result<Self::Reader, HardenedSha3Error> {
                 Self::finalize_xof_erasing_source(self)
             }